DEV: Modernize + adopt `PostList` component (#139)
This update makes use of core's new `PostList` component (https://github.com/discourse/discourse/pull/30604) to show the follow post feed. Additionally, it modernizes the route/controller around calling this component to make use of `RouteTemplate` instead of a route, template, and controller.

Tests already exist, but we update classes used in the test to be more fitting for the new component.